+++ lib/extensions/context_extensions.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 
+import '../models/platform_adaptive_icons.dart';
+
 /// Extensions on [BuildContext] for resolving platform-specific values.
 ///
 /// Provides convenience methods to select between Material (Android) and
@@ -8,6 +10,9 @@ import 'package:flutter/widgets.dart';
 // Purely for name-spacing collisions
 // ignore: prefer-match-file-name
 extension PlatformAdaptiveContextExtensions on BuildContext {
+  /// Render either a Material or Cupertino icon based on the platform
+  PlatformAdaptiveIcons get platformAdaptiveIcons => PlatformAdaptiveIcons(this);
+
   /// Returns the platform-appropriate [IconData].
   ///
   /// Selects [material] on Android and [cupertino] on iOS.