+++ lib/models/platform_adaptive_icons.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/widgets.dart';
+
+/// A class that provides adaptive icons based on the current platform.
+final class PlatformAdaptiveIcons {
+  /// The build context.
+  final BuildContext context;
+
+  /// Creates a new instance of [PlatformAdaptiveIcons].
+  const PlatformAdaptiveIcons(this.context);
+}